requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
joblib==0.17.0
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
